# exar-webui/openapi.yaml
openapi: 3.0.3
info:
  title: Daily Expenses API
  description: |
    REST API for the Daily Expenses application. Tracks shared spending for two
    people (Armin and Ramin), shops, catalog items, and invoices with line items.

    All monetary values are in USD. Dates in the API use ISO 8601 Gregorian
    calendar dates (`YYYY-MM-DD`). The web UI displays Jalali (Persian) dates.
  version: 1.0.0
  contact:
    name: Daily Expenses
servers:
  - url: http://localhost:8080/api
    description: Local development (default `ADDR=:8080`)
  - url: /api
    description: Same host as the web app (relative)
...
paths:
  /persons:
    get:
      tags: [Persons]
      summary: List persons
      description: Returns the two fixed persons (Armin and Ramin).
      operationId: listPersons
      responses:
        "200":
          description: List of persons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Person"
              example:
                - id: 1
                  name: Armin
                - id: 2
                  name: Ramin
        "500":
          $ref: "#/components/responses/InternalServerError"
  /shops:
    get:
      tags: [Shops]
      summary: List shops
      description: Returns all shops ordered by name.
      operationId: listShops
      responses:
        "200":
          description: List of shops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shop"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: [Shops]
      summary: Create shop
      description: Adds a new shop. Names must be unique.
      operationId: createShop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShopRequest"
            example:
              name: Grocery Store
      responses:
        "201":
          description: Shop created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Shop name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: shop already exists
        "500":
          $ref: "#/components/responses/InternalServerError"
  /shops/{id}:
    put:
      tags: [Shops]
      summary: Update shop
      description: Renames a shop by ID. Names must be unique.
      operationId: updateShop
      parameters:
        - $ref: "#/components/parameters/ShopId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShopRequest"
            example:
              name: Corner Market
      responses:
        "200":
          description: Shop updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Shop name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: shop already exists
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags: [Shops]
      summary: Delete shop
      description: |
        Deletes a shop by ID. Fails with 409 if any invoice references the shop.
      operationId: deleteShop
      parameters:
        - $ref: "#/components/parameters/ShopId"
      responses:
        "204":
          description: Shop deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Shop is referenced by one or more invoices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: shop is used by invoices
        "500":
          $ref: "#/components/responses/InternalServerError"
  /items:
    get:
      tags: [Items]
      summary: List or search items
      description: |
        Returns catalog items ordered by name. When `q` is provided with fewer
        than 3 characters, an empty array is returned. With 3+ characters,
        results are filtered with a case-insensitive substring match.
      operationId: listItems
      parameters:
        - name: q
          in: query
          description: Search query (min 3 characters to filter)
          schema:
            type: string
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags: [Items]
      summary: Create item
      description: Adds a catalog item name. Names must be unique.
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemRequest"
            example:
              name: Milk
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Item name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: item already exists
        "500":
          $ref: "#/components/responses/InternalServerError"
  /items/{id}:
    put:
      tags: [Items]
      summary: Update item
      description: Renames a catalog item by ID. Names must be unique.
      operationId: updateItem
      parameters:
        - $ref: "#/components/parameters/ItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemRequest"
            example:
              name: Organic Milk
      responses:
        "200":
          description: Item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Item name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: item already exists
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags: [Items]
      summary: Delete item
      description: Deletes a catalog item by ID. Always allowed (no invoice FK).
      operationId: deleteItem
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "204":
          description: Item deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /stats:
    get:
      tags: [Stats]
      summary: Get spend statistics
      description: |
        Returns total spend, share-weighted totals by person, and totals by shop.
        Optional Gregorian date range filters apply to all aggregates. Person
        totals use `invoice_total * share` for each person.
      operationId: getStats
      parameters:
        - name: from_date
          in: query
          description: Include invoices on or after this date (`YYYY-MM-DD`)
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          description: Include invoices on or before this date (`YYYY-MM-DD`)
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
    ShopId:
      name: id
      in: path
      required: true
      description: Shop ID
      schema:
        type: integer
        minimum: 1
    ItemId:
      name: id
      in: path
      required: true
      description: Item ID
      schema:
        type: integer
        minimum: 1
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Person:
      type: object
      properties:
        id:
          type: integer
          description: Person ID (1 or 2)
          enum: [1, 2]
        name:
          type: string
      required: [id, name]
    Shop:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required: [id, name]
    CreateShopRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
      required: [name]
    UpdateShopRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
      required: [name]
    Item:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required: [id, name]
    CreateItemRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
      required: [name]
    UpdateItemRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
      required: [name]
    PersonTotal:
      type: object
      properties:
        person_id:
          type: integer
          enum: [1, 2]
        person_name:
          type: string
        total:
          type: number
          format: double
      required: [person_id, person_name, total]
    ShopTotal:
      type: object
      properties:
        shop_id:
          type: integer
        shop_name:
          type: string
        total:
          type: number
          format: double
      required: [shop_id, shop_name, total]
    Stats:
      type: object
      properties:
        total:
          type: number
          format: double
        by_person:
          type: array
          items:
            $ref: "#/components/schemas/PersonTotal"
        by_shop:
          type: array
          items:
            $ref: "#/components/schemas/ShopTotal"
      required: [total, by_person, by_shop]
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"